// src/api/model/createTaxFilingDtoSubmissionType.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * accounting
 * Accounting API description
 * OpenAPI spec version: 1.0
 */

export type CreateTaxFilingDtoSubmissionType = typeof CreateTaxFilingDtoSubmissionType[keyof typeof CreateTaxFilingDtoSubmissionType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const CreateTaxFilingDtoSubmissionType = {
  REGULAR: 'REGULAR',
  CORRECTIVE: 'CORRECTIVE',
  FOLLOWUP: 'FOLLOWUP',
} as const;