// src/api/model/invoiceResponseDtoKind.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * accounting
 * Accounting API description
 * OpenAPI spec version: 1.0
 */

export type InvoiceResponseDtoKind = typeof InvoiceResponseDtoKind[keyof typeof InvoiceResponseDtoKind];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const InvoiceResponseDtoKind = {
  INVOICE: 'INVOICE',
  SIMPLE: 'SIMPLE',
} as const;